Reorganize files
Source and header files have been reorganized into 2 categories core and helper.
Core is the main parser/interpreter and helper are some functions that make life just a little bit easier.
This resolves issue #9
