+++ core/aml_opcodes.h
