+++ core/exec.c
