+++ core/exec2.c
