+++ core/libc.c
