+++ core/libc.h
