+++ core/ns.c
