+++ core/ns_impl.h
