+++ core/opregion.c
