+++ core/opregion.h
