+++ core/os_methods.c
