+++ core/util-hash.h
