+++ helpers/resource.c
@@ -8,8 +8,8 @@
 /* Allows discovering of each device's used resources, and thus is needed
  * for basic system enumeration as well as PCI IRQ routing. */
 
-#include <lai/core.h>
-#include "libc.h"
+#include <lai/helpers/resource.h>
+#include "../core/libc.h"
 
 #define ACPI_SMALL_IRQ            0x04
 #define ACPI_SMALL_DMA            0x05
