+++ include/lai/helpers/resource.h
@@ -1,0 +1,10 @@
+/*
+ * Lux ACPI Implementation
+ * Copyright (C) 2018-2019 the lai authors
+ */
+
+#pragma once
+
+#include <lai/core.h>
+
+size_t lai_read_resource(lai_nsnode_t *, acpi_resource_t *);
