+++ include/lai/helpers/sleep.h
@@ -1,0 +1,10 @@
+/*
+ * Lux ACPI Implementation
+ * Copyright (C) 2018-2019 the lai authors
+ */
+
+#pragma once
+
+#include <lai/core.h>
+
+int lai_enter_sleep(uint8_t);
