+++ include/lai/internal-ns.h
@@ -1,6 +1,6 @@
 /*
  * Lux ACPI Implementation
- * Copyright (C) 2018-2019 by Omar Muhamed
+ * Copyright (C) 2018-2019 the lai authors
  */
 
 #pragma once
