+++ meson.build
@@ -4,18 +4,18 @@ project('lai', 'c')
 include = include_directories('include')
 
 library = static_library('lai',
-        'src/eval.c',
-        'src/exec.c',
-        'src/exec2.c',
-        'src/libc.c',
-        'src/ns.c',
-        'src/opregion.c',
-        'src/os_methods.c',
-        'src/pciroute.c',
-        'src/resource.c',
-        'src/sci.c',
-        'src/sleep.c',
-        'src/vsnprintf.c',
+        'core/eval.c',
+        'core/exec.c',
+        'core/exec2.c',
+        'core/libc.c',
+        'core/ns.c',
+        'core/opregion.c',
+        'core/os_methods.c',
+        'helpers/pciroute.c',
+        'helpers/resource.c',
+        'helpers/sci.c',
+        'helpers/sleep.c',
+        'core/vsnprintf.c',
     include_directories: include,
     pic: false)
